Move lib/ to kitsune/lib
