+++ kitsune/lib/sumo_locales.py
