+++ wsgi/kitsune.wsgi
@@ -26,7 +26,7 @@ site.addsitedir(os.path.abspath(os.path.join(wsgidir, '../')))
 # For django-celery
 os.environ['CELERY_LOADER'] = 'django'
 
-# manage adds /apps, /lib, and /vendor to the Python path.
+# manage /vendor to the Python path.
 import manage
 
 import django.conf
